+++ src/Tail.java
@@ -8,8 +8,8 @@ import java.util.LinkedList;
  */
 public class Tail extends Circle
 {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
     public Direction direction;
     private Queue<Move> nextMove; //the next moves this tail is going to make
     
@@ -47,7 +47,7 @@ public class Tail extends Circle
         setCenterY(y * 20 + 10);
     }
 
-    public Tail(Queue<Move> nextMove, int x, int y, Direction direction)
+    public Tail(Queue<Move> nextMove, double x, double y, Direction direction)
     {
         // circle properties
         super(8);
@@ -83,7 +83,7 @@ public class Tail extends Circle
         return "x: "+ x + " y: " + y + " direction: " + direction;
     }
     
-    public void updateFrame(int movement)
+    public void updateFrame(double movement)
     {
         
         if(nextMove.peek() != null) {
@@ -149,22 +149,22 @@ public class Tail extends Circle
         return nextMove;
     }
     
-    public int getX()
+    public double getX()
     {
         return x;
     }
     
-    public void setX(int x)
+    public void setX(double x)
     {
         this.x = x;
     }
     
-    public int getY()
+    public double getY()
     {
         return y;
     }
     
-    public void setY(int y)
+    public void setY(double y)
     {
         this.y = y;
     }
